Changes that lead to the development of the new CPS formulae engine Add getters "len" and "rad" for CPS enhancement experiments
Also whitespace got auto-corrected

Also the FIXME comment for the new properties is important:

We should have proper whitelisting facilities in CPS
to overcome 'accidential' access to properties that weren't meant
to be used in CPS. Also, this could add a re-mapping of names
to getter methods (or getters). This would make the namespace
less crowded and allow a better names for he CPS user!

a comment typo and whitespace

add operators "polar" and "deg"

This is the experimental beginning of an CPS enhancement that will let us
describe point positions as a distance plus an angle. This is an important part
to make our pen model more complete.

algebra with numbers in CompoundVector context, deg operator is now before its argument

also:
added some missing semicolons
a better error message, indicating the used operator by its literal name

for now and experimental removal of the implicit intrinsic value in compound vectors

whitespace and a typo in a comment

add some facilities to process Vector data types

This is a preparation to unify the code of CompundReal and CompoundVector
in the end. This means there is code duplication from CompoundVector and
this is intentional but temporary.

Change inDir(Intrinsic) and outDir(Intrinsic) to be given as radian (instead of a vector

This will help to keep the right directions when transforming the skeleton.

Also, some automatical whitespace conversions happend.

add parameters 'onLength' and 'onDir' to enable more pen control

use onLength and onDir to construct the offset of left and right on curve points from center

Also use the inverse of onDir of right to specify onDir of left
and use onLength of right directly for onLength of left

import right.onLength and right.onDir

default setup of onDir and onLength of point>left

add a simple scale operator to compound vector

delay querying some values unless the tensions have been verified

copy and pasted the subtract command of compoundVectorFactory

Added index command to access numeric properties

Currently used to access x at 0 and y at 1 of a vector, which makes
some actions in cps possible.

This is a quick and dirty implementation, it will be replaced by a
rewrite of the compound*CPSFactory classes