Add code to handle cache invalidation on OMA changes correctly.
Add a 'subtree' key to the OMA onPropertyChange channel. Which is
a virtual property (can't be read from the _Node directly). This
is used to invalidate CPS caches that are keyed with a SelectorList.
I.e. The result of running a node.query from CPS.

Add a new strategy to StyleDict, to invalidate the caches, paralell
to `styleDict.invalidateRules`: `styleDict.checkRules`. It should
be measured which one is faster though. Going with `styleDict.checkRules`
for now.

Add OMA onPropertyChange keys for 'children', 'id', 'index' and the
virtual property 'tail-index' which are used to handle OMA manipulaton
related cache invalidations.

Add a `child -> child index` map, to faster read children indexes from
a parent.