Implement OMA Node-Id management
Until now the uniqueness rules of Node Ids were nowhere enforced nor controlled.
This adds flexible control for Ids and their uniqueness in OMA based models.
The default is as in XML a document wide ID-uniqueness, but it is also straight
forward to implement a id-uniqueness model as we find it in MOM (which was
inherited by the ruls of unique identifiers in UFO) and makes totally sense.

In the case of MOM uniqueness is organized per Glyph and all it's descendants.
Then one level above a Master only manages the uniqueness of Glyph IDs (=Glyph
names). Technically a node stops managing ids in a subtree if the parent element
of the subtree manages its descendants itself.